Fix issues in ClubController

+++ spring-boot/clubMicroservice/src/main/java/com/example/clubMicroservice/functions/RequestToClubFunction.java
@@ -13,7 +13,7 @@ public class RequestToClubFunction implements Function<PostClubRequest, Club> {
     public Club apply(PostClubRequest postClubRequest){
         return Club.builder()
                 .name(postClubRequest.getName())
-                .yearOfFoundation(postClubRequest.getYearOfFoundation())
+                .yearOfFoundation(postClubRequest.getYearOfFoundation().intValue())
                 .ground(postClubRequest.getGround())
                 .city(postClubRequest.getCity())
                 .nickname(postClubRequest.getNickname())
